pyproject.toml
[project]
name = "graphafold"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "ipykernel>=6.29.5",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "ruff>=0.11.6",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "tool>=0.8.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
